gamey: refactorizar preámbulo común de bots. #171
Extraer preámbulo común a módulo bot_helper.
